Update Paryments/createOrder and chat/getResponse
Refactor createOrder controller for improved code maintenance and readability, and enhance chat/getResponse controller to handle expanded configuration options.
